重构 CardGen
1. 把生成事件卡和生成非事件卡分开为两个程序。
2. 主要重构非事件卡部分
3. 做了一些数据和逻辑的分离。
4. 可读性改进。
