+++ src/ColorMap.py
@@ -1,0 +1,24 @@
+COLOR_MAP = {
+    "standardWhiteColor": (251, 240, 217),
+    "standardBlackColor": (65, 65, 65),
+    "化学": {
+        "classTextColor": (29, 48, 161),
+        "descriptionColor": (125, 139, 151),
+    },
+    "机械": {"classTextColor": (103, 64, 29), "descriptionColor": (153, 138, 11)},
+    "巫术": {"classTextColor": (78, 19, 102), "descriptionColor": (124, 104, 125)},
+    "自然": {
+        "classTextColor": (28, 75, 24),
+        "descriptionColor": (80, 106, 54),
+    },
+    "神秘": {
+        "classTextColor": (54, 54, 54),
+        "descriptionColor": (118, 123, 129),
+    },
+    "通用": {
+        "classTextColor": (90, 90, 90),
+        "descriptionColor": (128, 139, 151),
+    },
+}
+
+
